Duplicate zeppelin tests logic to test the WBTC token.
As in zeppelin tests, the different test behaviors are separated to different
files.

In order to examine differences with zeppelin tests one can diff the behaviour
files with those of zeppelin (under node_modules/open_zeppelin/tests/..).

Importing (with require) the actual zeppelin tests is impossible since some of
their tests are without separation to behavior and test token.
Also as some of the token logic is overriden in wbtc, some of their test
logic should be changed anyway.

Also added mock contracts used for these tests, as zeppelin npm does not
include the mocks and their testing scripts.