Remove inheritance of HasNoEther in OwnableContract.
This is done because the ownableContractOwner's callReclaimEther() function is
unuseable anyhow, since ownableCOntractOwner does not accept ether.
So no real need for callReclaimEther.

It Should be noted that in any case all contracts that had the HasNoEther
attribute do not have a default payable function, so in any case they can't
recieve ether (except from self-destruct and mining to them).