Align all non pure/view functions to return bool.
Behaviour for such functions is to return true or revert.

An exeption for this alignment work is the token iplementation functions,
where zeppelin implementation is not consistent and does not return a bool.
Overriding functions in WBTC.sol keep the same return value as in zeppelin
overriden functions, even though inconsistent in terms of return values.
In such cases (e.g burn, pause, unpause), a return value is added in the
calling/wrapping function in Controller.sol.