Align deployment script to recent changes, add deployment script test.
This will support continous integration of the deployment script.

In order to support testing the deployment script in truffle, it was split to
an internal package (scripts/deployerImplementation.js) and a wrapper
(scripts/deployer.js).
The internal package is used both by the truffle test and by the wrapper.
The wrapper is used for running the script from command line.