More functions ordering, parent ctr call style change, other neat stuff.
Change order of some functions in Members to align with a fixed
(chronological minded) pattern of:
-first custodian related functions, than merchant related functions.

Change order of some functions in Members to align with a fixed
(chronological minded) pattern of:
-first mint related functions, than burn related functions.
-an exception to the above is that setCustodianBtcDepositAddress and
setMerchantBtcDepositAddress are together at the beginning (as they are done
before everything).

Change wbtc call to DetailedERC20 constructor to be part of the contract
declaration and thus avoid an solhint no-empty-block comment.

Change to single space between each contract's pragma and imports
(as in zeppelin code).

Make sure each contract file ends wih a line break.